+++ benches/util/mod.rs
@@ -1,4 +1,5 @@
-use bevy_math::UVec2;
+use bevy_math::{uvec2, UVec2};
+use image::{DynamicImage, GenericImageView, Rgba};
 use pixel_map::PixelMap;
 
 pub fn create_checker_board(size: &UVec2) -> PixelMap {
@@ -9,4 +10,19 @@ pub fn create_checker_board(size: &UVec2) -> PixelMap {
         }
     }
     pixel_map
+}
+
+pub fn load_image(pixel_map: &mut PixelMap<Rgba<u8>>, image: &DynamicImage) {
+    let region_width = image.width();
+    let region_height = image.height();
+
+    for y in 0..region_height {
+        for x in 0..region_width {
+            let image_x = x;
+            let image_y = y;
+            let color = image.get_pixel(image_x, image_y);
+            let point = uvec2(x, region_height - y - 1);
+            pixel_map.set_pixel(point, color);
+        }
+    }
 }